remove previous tryouts
