+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/create-runtime.ts
@@ -138,14 +138,6 @@ ${injectCode("./extract-locale-from-url-v2.js")}
 
 ${injectCode("./localize-url.js")}
 
-${injectCode("./localize-url-v2.js")}
-
-${injectCode("./localize-url-v3.js")}
-
-${injectCode("./localize-url-v4.js")}
-
-${injectCode("./de-localize-url.js")}
-
 // ------ TYPES ------
 
 /**
