+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/localize-url.test.ts
@@ -1,89 +1,6 @@
 import { test, expect } from "vitest";
 import { createRuntimeForTesting } from "./create-runtime.js";
 
-test("matches according to the order", async () => {
-	const runtime = await createRuntimeForTesting({
-		baseLocale: "en",
-		locales: ["en", "de", "fr", "en-UK"],
-		compilerOptions: {
-			urlPatterns: [
-				{
-					pattern: `http{s}\\://*host.co.uk{/de/*path}`,
-					locale: "de",
-					deLocalizedPattern: "http{s}\\://*host.co.uk{/*path}",
-				},
-				{
-					pattern: `http{s}\\://*host.co.uk{/*path}`,
-					locale: "en-UK",
-					deLocalizedPattern: "http{s}\\://*host.co.uk{/*path}",
-				},
-				// default pathname has locale pattern
-				{
-					pattern: "http{s}\\://*host.com/fr{/*path}",
-					locale: "fr",
-					deLocalizedPattern: "http{s}\\://*host.com/{*path}",
-				},
-				{
-					pattern: "http{s}\\://*host.com/de{/*path}",
-					locale: "de",
-					deLocalizedPattern: "http{s}\\://*host.com/{*path}",
-				},
-				{
-					pattern: "http{s}\\://*host.com/{*path}",
-					locale: "en",
-					deLocalizedPattern: "http{s}\\://*host.com/{*path}",
-				},
-			],
-		},
-	});
-
-	expect(
-		runtime.localizeUrl("https://example.co.uk/about", { locale: "en-UK" })
-	).toBe("https://example.co.uk/about");
-	expect(
-		runtime.localizeUrl("https://example.co.uk/about", { locale: "de" })
-	).toBe("https://example.co.uk/de/about");
-	expect(
-		runtime.localizeUrl("https://example.co.uk/about", { locale: "fr" })
-	).toBe("https://example.com/fr/about");
-	expect(
-		runtime.localizeUrl("https://example.co.uk/about", { locale: "en" })
-	).toBe("https://example.com/about");
-});
-
-test("can handle relative urls", async () => {
-	const runtime = await createRuntimeForTesting({
-		baseLocale: "en",
-		locales: ["en", "de", "fr", "en-UK"],
-		compilerOptions: {
-			urlPatterns: [
-				{
-					pattern: `http{s}\\://*host.co.uk{/de/*path}`,
-					locale: "de",
-					deLocalizedPattern: "http{s}\\://*host.co.uk{/*path}",
-				},
-				{
-					pattern: `http{s}\\://*host.co.uk{/*path}`,
-					locale: "en",
-					deLocalizedPattern: "http{s}\\://*host.co.uk{/*path}",
-				},
-			],
-		},
-	});
-
-	globalThis.window = {
-		location: {
-			origin: "https://example.co.uk",
-		},
-	};
-	expect(runtime.localizeUrl("/about", { locale: "de" })).toBe(
-		"https://example.co.uk/de/about"
-	);
-	expect(runtime.localizeUrl("/about", { locale: "en" })).toBe(
-		"https://example.co.uk/about"
-	);
-});
-
 test("handles translated path segments", async () => {
 	const runtime = await createRuntimeForTesting({
 		baseLocale: "en",
@@ -91,14 +8,11 @@ test("handles translated path segments", async () => {
 		compilerOptions: {
 			urlPatterns: [
 				{
-					pattern: `*protocol\\://*domain/buchladen{/:id}`,
-					locale: "de",
-					deLocalizedPattern: "*protocol\\://*domain/bookstore{/:id}",
-				},
-				{
-					pattern: `*protocol\\://*domain/bookstore{/:id}`,
-					locale: "en",
-					deLocalizedPattern: "*protocol\\://*domain/bookstore{/:id}",
+					pattern: "https://:domain.:tld/:bookstore/:path*",
+					localizedParams: {
+						en: { bookstore: "bookstore" },
+						de: { bookstore: "buchladen" },
+					},
 				},
 			],
 		},
@@ -118,83 +32,95 @@ test("handles translated path segments", async () => {
 	expect(
 		runtime.localizeUrl("https://example.com/bookstore/45", { locale: "de" })
 	).toBe("https://example.com/buchladen/45");
+
+	// delocalizing
+	expect(runtime.deLocalizeUrl("https://example.com/buchladen/45")).toBe(
+		"https://example.com/bookstore/45"
+	);
 });
 
-test("can handle pathname matching", async () => {
+test("cross domain urls", async () => {
 	const runtime = await createRuntimeForTesting({
 		baseLocale: "en",
-		locales: ["en", "de", "fr", "en-UK"],
+		locales: ["en", "de"],
 		compilerOptions: {
 			urlPatterns: [
 				{
-					pattern: `/de/{*path}`,
-					locale: "de",
-					deLocalizedPattern: "{/*path}",
-				},
-				{
-					pattern: `/{*path}`,
-					locale: "en",
-					deLocalizedPattern: "{/*path}",
+					pattern: "https://:domain(.*)/:path*",
+					localizedParams: {
+						en: { domain: "example.com" },
+						de: { domain: "de.example.com" },
+					},
 				},
 			],
 		},
 	});
 
-	expect(runtime.localizeUrl("/about", { locale: "de" })).toBe("/de/about");
-	expect(runtime.localizeUrl("/about", { locale: "en" })).toBe("/about");
-});
-
-test("can handle domain based matching", async () => {
-	const runtime = await createRuntimeForTesting({
-		baseLocale: "en",
-		locales: ["en", "de", "fr"],
-		compilerOptions: {
-			urlPatterns: [
-				{
-					// 1) Domain-based, root locale is "de"
-					base: "http{s}://de.example.com",
-					rootLocale: "de",
-					// Other locales can be path-based or domain-based
-					locales: {
-						de: "/{*path}", // "de" is root → just "/about"
-						en: "http{s}://example.com/{*path}", // Switch to domain-based for English
-						fr: "http{s}://fr.example.com/{*path}",
-					},
-				},
-				{
-					// 2) Domain-based, root locale is "en"
-					base: "http{s}://example.com",
-					rootLocale: "en",
-					locales: {
-						en: "/{*path}", // "en" is root → just "/about"
-						de: "/de/{*path}", // path-based for German
-						fr: "http{s}://fr.example.com/{*path}", // cross-domain for French
-					},
-				},
-				{
-					// 3) Domain-based, root locale is "fr"
-					base: "http{s}://fr.example.com",
-					rootLocale: "fr",
-					locales: {
-						fr: "/{*path}", // "fr" root → just "/about"
-						en: "http{s}://example.com/{*path}",
-						de: "http{s}://de.example.com/{*path}",
-					},
-				},
-			],
-		},
-	});
+	// DE routing
+	expect(
+		runtime.localizeUrl("https://de.example.com/about", { locale: "de" })
+	).toBe("https://de.example.com/about");
 
 	expect(
-		runtime.localizeUrl("https://example.de/about", { locale: "de" })
-	).toBe("https://example.de/about");
+		runtime.localizeUrl("https://de.example.com/about", { locale: "en" })
+	).toBe("https://example.com/about");
+
+	// delocalizing
+	expect(runtime.deLocalizeUrl("https://de.example.com/about")).toBe(
+		"https://example.com/about"
+	);
+
+	// EN routing
+	expect(
+		runtime.localizeUrl("https://example.com/about", { locale: "de" })
+	).toBe("https://de.example.com/about");
+
 	expect(
 		runtime.localizeUrl("https://example.com/about", { locale: "en" })
 	).toBe("https://example.com/about");
-	// simulating shared domains
+
+	// delocalizing
+	expect(runtime.deLocalizeUrl("https://example.com/about")).toBe(
+		"https://example.com/about"
+	);
+});
+
+test("pathname based localization", async () => {
+	const runtime = await createRuntimeForTesting({
+		baseLocale: "en",
+		locales: ["en", "de"],
+		compilerOptions: {
+			urlPatterns: [
+				{
+					pattern: "https://:domain(.*)/:locale(de)?/:path*",
+					localizedParams: {
+						de: { locale: "de" },
+						en: { locale: null },
+					},
+				},
+			],
+		},
+	});
+
+	// en to de
 	expect(
-		runtime.localizeUrl("https://example.com/about", { locale: "fr" })
-	).toBe("https://example.com/fr/about");
+		runtime.localizeUrl("https://example.com/about", { locale: "de" })
+	).toBe("https://example.com/de/about");
+
+	// de to en
+	expect(
+		runtime.localizeUrl("https://example.com/de/about", { locale: "en" })
+	).toBe("https://example.com/about");
+
+	// en == delocalized
+	expect(
+		runtime.localizeUrl("https://example.com/about", { locale: "en" })
+	).toBe("https://example.com/about");
+
+	// delocalizing
+	expect(runtime.deLocalizeUrl("https://example.com/de/about")).toBe(
+		"https://example.com/about"
+	);
 });
 
 test("multi tenancy", async () => {
@@ -203,45 +129,62 @@ test("multi tenancy", async () => {
 		locales: ["en", "de", "fr"],
 		compilerOptions: {
 			urlPatterns: [
+				// 1) customer1.fr => root locale is fr, sub-locale is /en/
 				{
-					// customer1 has french as root locale
-					pattern: `http{s}\\://customer1.fr{/*path}`,
-					locale: "fr",
-					deLocalizedPattern: "http{s}\\://customer1.fr{/*path}",
+					pattern: "https://customer1.fr/:locale(en)?/:path*",
+					localizedParams: {
+						fr: { locale: null }, // remove /locale => root is FR
+						en: { locale: "en" }, // subpath => /en/about
+					},
+					deLocalizedParams: { locale: null },
 				},
+				// 2) customer2.com => root locale is en, sub-locale is /fr/
 				{
-					// customer2 has english as root locale
-					pattern: `http{s}\\://customer2.com{/*path}`,
-					locale: "en",
-					deLocalizedPattern: "http{s}\\://customer2.com{/*path}",
+					pattern: "https://customer2.com/:locale(fr)?/:path*",
+					localizedParams: {
+						en: { locale: null }, // remove /locale => root is EN
+						fr: { locale: "fr" }, // subpath => /fr/about
+					},
+					deLocalizedParams: { locale: "fr" },
 				},
+				// 3) Any other domain => path-based for en/fr
 				{
-					// both customers use prefixed locales for their non root locale
-					pattern: `http{s}\\://*domain/en{/*path}`,
-					locale: "en",
-					deLocalizedPattern: "http{s}\\://*domain/en{/*path}",
-				},
-				{
-					// both customers use prefixed locales for their non root locale
-					pattern: `http{s}\\://*domain/fr{/*path}`,
-					locale: "fr",
-					deLocalizedPattern: "http{s}\\://*domain/fr{/*path}",
+					pattern: "https://:domain(.*)/:locale/:path*",
+					localizedParams: {
+						en: { locale: "en" },
+						fr: { locale: "fr" },
+					},
 				},
 			],
 		},
 	});
-	// customer 1 - french locale is root
+	// customer 1 - localizing french to french
 	expect(
 		runtime.localizeUrl("https://customer1.fr/about", { locale: "fr" })
 	).toBe("https://customer1.fr/about");
 
-	// customer 1 - english locale is subpath
+	// customer 1 - localizing from french to english
 	expect(
 		runtime.localizeUrl("https://customer1.fr/about", { locale: "en" })
 	).toBe("https://customer1.fr/en/about");
 
-	// customer 2 - english locale is root
+	// customer 1 - de-localizing french
+	expect(runtime.deLocalizeUrl("https://customer1.fr/en/about")).toBe(
+		"https://customer1.fr/about"
+	);
+
+	// customer 2 - english to english
 	expect(
-		runtime.localizeUrl("https://customer2.fr/about", { locale: "en" })
-	).toBe("https://customer2.fr/about");
+		runtime.localizeUrl("https://customer2.com/about", { locale: "en" })
+	).toBe("https://customer2.com/about");
+
+	// customer 2 - english to french
+	expect(
+		runtime.localizeUrl("https://customer2.com/about", { locale: "fr" })
+	).toBe("https://customer2.com/fr/about");
+
+	// customer 2 - de-localize french
+	expect(runtime.deLocalizeUrl("https://customer2.com/about")).toBe(
+		"https://customer2.com/fr/about"
+	);
 });
